Feat: Added extra method
Two extra methods are added in batch

+++ ComputerSeekho/src/main/java/com/Project/Services/BatchServiceImpl.java
@@ -9,6 +9,8 @@ import org.springframework.stereotype.Service;
 import com.Project.Entities.Batch;
 import com.Project.Repositories.BatchRepository;
 
+import jakarta.transaction.Transactional;
+
 @Service
 public class BatchServiceImpl implements BatchService{
 
@@ -16,28 +18,47 @@ public class BatchServiceImpl implements BatchService{
     BatchRepository batchRepository;
 
     @Override
-    public void addBatch(Batch b) {
-        batchRepository.save(b);
+    public Batch addBatch(Batch b) {
+       return batchRepository.save(b);
     }
-
     @Override
     public List<Batch> getAllBatches() {
         return batchRepository.findAll();
     }
-
     @Override
     public Optional<Batch> getByBatchName(String batch_name) {
         return batchRepository.findByBatchName(batch_name);
     }
-
     @Override
-    public void delete(int batch_id) {
-        batchRepository.deleteById(batch_id);
+    public boolean delete(int batch_id) {
+
+    Batch delete= batchRepository.findById(batch_id).get();
+    if(delete==null)
+    {
+        return false;
     }
-
+    batchRepository.deleteById(batch_id);
+    return true;
+    }
     @Override
-    public void activateBatch(int batch_id,Boolean batch_is_active) {
-        batchRepository.activateBatch(batch_is_active,batch_id);
+    @Transactional
+    public int deactivateBatch(int batch_id) {
+        Optional<Batch> batchOptional = batchRepository.findById(batch_id);
+        if (batchOptional.isPresent()) {
+            batchRepository.updateBatchIsActive(batch_id);
+            return 1;
+             
+        } else {
+            return 0;
+        }
+    }
+    @Override
+    public List<Batch> getAllActiveBatches() {
+        return batchRepository.findAllActiveBatch();
+    }
+    @Override
+    public List<Batch> getByCourseId(int course_id) {
+        return batchRepository.findByCourseId(course_id);
     }
     
 }
